Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/webrtc/RTCPeerConnection-onsignalingstatechanged.https-expected.txt


PASS Negotiation methods fire signalingstatechange events
FAIL Closing a PeerConnection should not fire signalingstatechange event assert_unreached: Reached unreachable code
PASS signalingstatechange is the first event to fire
